task1 и описание в файле function.js ver2
