+++ js/functions.js
@@ -1,30 +1,21 @@
-//     Функция для проверки длины строки. Она принимает строку, которую нужно проверить, и максимальную длину и возвращает true, если строка меньше или равна указанной длине, и false, если строка длиннее.
-
+/**
+ *      Функция для проверки длины строки. Она принимает строку, которую нужно проверить, и максимальную длину и возвращает true, если строка меньше или равна указанной длине, и false, если строка длиннее.
+ * @param {int} maxLength - длина строки для проверки
+ * @param {string} someString - какая-то строка
+ * @return {boolean} - истина, если строка меньше либо равна заданой длине
+ */
 function lengthCheck (someString, maxLength) {
   return maxLength <= someString.length;
 }
 
-lengthCheck('12345', 5);
-
-
-//     Функция для проверки, является ли строка палиндромом.
-
-// мне нравитсья эта реализация, но по заданию надо сделать не так
-/*
+/**
+ * Функция для проверки, является ли строка палиндромом.
+ * @param {string} polindrome - проверяемая строка
+ * @returns {boolean} - истина, если полиндром
+ */
 function polindromeTrue (polindrome) {
   return polindrome.toUpperCase().replaceAll(' ','') === polindrome.toUpperCase().replaceAll(' ','').split('').reverse().join('');
 }
-*/
 
-function polindromeTrue (polindrome) {
-  const oldString = polindrome.toUpperCase().replaceAll(' ','');
-  let newString = '';
-  for (let i = oldString.length - 1; i >= 0; i--) {
-    newString += oldString.at(i);
-  }
-  return oldString === newString;
-
-}
-
-polindromeTrue('Лёша на полке клопа нашёл ');
+export {polindromeTrue, lengthCheck};
